+++ handnew04/app/src/main/java/com/example/handnew04/MainApplication.kt
@@ -1,7 +1,11 @@
 package com.example.handnew04
 
 import android.app.Application
-import com.example.handnew04.di.*
+import com.example.handnew04.di.mainViewModelModule
+import com.example.model.di.localDataModule
+import com.example.model.di.movieRepositoryModule
+import com.example.model.di.networkManagerModule
+import com.example.model.di.remoteDataModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
@@ -11,7 +15,7 @@ class MainApplication : Application() {
         super.onCreate()
         startKoin {
             androidLogger()
-            androidContext(this@MainApplication)
+            androidContext(this@MainApplication.applicationContext)
             modules(
                 listOf(
                     localDataModule,
