+++ handnew04/app/src/main/java/com/example/handnew04/adapter/DataBindingAdapter.kt
@@ -9,7 +9,7 @@ import android.widget.TextView
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.example.handnew04.data.MovieData
+import com.example.model.data.MovieData
 
 @BindingAdapter("bind:bindImage")
 fun bindImage(imageView: ImageView, imageUri: String) {
