--- handnew04/app/src/main/java/com/example/handnew04/data/MovieDataResponse.kt
@@ -1,16 +1,0 @@
-package com.example.handnew04.data
-
-data class NaverMovieResponse(
-    var items: List<MovieData>
-)
-
-data class MovieData(
-    val title: String,
-    val link: String,
-    val image: String,
-    val pubDate: String,
-    val director: String,
-    val actor: String,
-    val userRating: String
-)
-
