--- handnew04/app/src/main/java/com/example/handnew04/data/MovieRepository.kt
@@ -1,16 +1,0 @@
-package com.example.handnew04.data
-
-import com.example.handnew04.data.local.MovieLocalDataSource
-import com.example.handnew04.data.remote.MovieRemoteDataSource
-
-class MovieRepository(
-    val movieLocalDataSource: MovieLocalDataSource,
-    val movieRemoteDataSource: MovieRemoteDataSource
-) {
-    fun getMovieData(
-        query: String,
-        success: (NaverMovieResponse) -> Unit,
-        fail: (Throwable) -> Unit
-    ) =
-        movieRemoteDataSource.getMovieData(query, success, fail)
-}
