--- handnew04/app/src/main/java/com/example/handnew04/data/local/MovieLocalDataSource.kt
@@ -1,5 +1,0 @@
-package com.example.handnew04.data.local
-
-interface MovieLocalDataSource {
-    fun getMovieData()
-}
