--- handnew04/app/src/main/java/com/example/handnew04/data/local/MovieLocalDataSourceImpl.kt
@@ -1,8 +1,0 @@
-package com.example.handnew04.data.local
-
-class MovieLocalDataSourceImpl : MovieLocalDataSource {
-    override fun getMovieData() {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-    }
-
-}
