--- handnew04/app/src/main/java/com/example/handnew04/data/remote/MovieRemoteDataSource.kt
@@ -1,11 +1,0 @@
-package com.example.handnew04.data.remote
-
-import com.example.handnew04.data.NaverMovieResponse
-
-interface MovieRemoteDataSource {
-    fun getMovieData(
-        query: String,
-        success: (NaverMovieResponse) -> Unit,
-        fail: (Throwable) -> Unit
-    )
-}
