--- handnew04/app/src/main/java/com/example/handnew04/di/LocalDataModule.kt
@@ -1,9 +1,0 @@
-package com.example.handnew04.di
-
-import com.example.handnew04.data.local.MovieLocalDataSource
-import com.example.handnew04.data.local.MovieLocalDataSourceImpl
-import org.koin.dsl.module
-
-val localDataModule = module {
-    single<MovieLocalDataSource> { MovieLocalDataSourceImpl() }
-}
