--- handnew04/app/src/main/java/com/example/handnew04/di/NetworkManagerModule.kt
@@ -1,8 +1,0 @@
-package com.example.handnew04.di
-
-import com.example.handnew04.network.NetworkManager
-import org.koin.android.ext.koin.androidApplication
-import org.koin.dsl.module
-
-val networkManagerModule =
-    module { single { NetworkManager(androidApplication()) } }
