--- handnew04/app/src/main/java/com/example/handnew04/di/RepositoryModule.kt
@@ -1,6 +1,0 @@
-package com.example.handnew04.di
-
-import com.example.handnew04.data.MovieRepository
-import org.koin.dsl.module
-
-val movieRepositoryModule = module { single { MovieRepository(get(), get()) } }
