+++ handnew04/app/src/main/java/com/example/handnew04/ui/main/MainViewModel.kt
@@ -3,9 +3,10 @@ package com.example.handnew04.ui.main
 import android.util.Log
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.example.handnew04.data.MovieData
-import com.example.handnew04.data.MovieRepository
-import com.example.handnew04.network.NetworkManager
+import com.example.model.data.MovieData
+import com.example.model.data.MovieRepository
+import com.example.model.network.NetworkManager
+
 
 class MainViewModel(val networkManager: NetworkManager, val movieRepository: MovieRepository) :
     ViewModel() {
