+++ handnew04/app/src/main/res/layout/item_movie.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="movie"
-            type="com.example.handnew04.data.MovieData" />
+            type="com.example.model.data.MovieData" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
