+++ handnew04/model/proguard-rules.pro
