+++ handnew04/model/src/main/java/com/example/model/data/MovieDataResponse.kt
@@ -1,0 +1,15 @@
+package com.example.model.data
+
+data class NaverMovieResponse(
+    var items: List<MovieData>
+)
+
+data class MovieData(
+    val title: String,
+    val link: String,
+    val image: String,
+    val pubDate: String,
+    val director: String,
+    val actor: String,
+    val userRating: String
+)
