+++ handnew04/model/src/main/java/com/example/model/data/local/MovieLocalDataSource.kt
@@ -1,0 +1,5 @@
+package com.example.model.data.local
+
+interface MovieLocalDataSource {
+    fun getMovieData()
+}
