+++ handnew04/model/src/main/java/com/example/model/data/local/MovieLocalDataSourceImpl.kt
@@ -1,0 +1,8 @@
+package com.example.model.data.local
+
+class MovieLocalDataSourceImpl : MovieLocalDataSource {
+    override fun getMovieData() {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+
+}
