+++ handnew04/model/src/main/java/com/example/model/data/remote/MovieRemoteDataSource.kt
@@ -1,0 +1,12 @@
+package com.example.model.data.remote
+
+import com.example.model.data.NaverMovieResponse
+
+
+interface MovieRemoteDataSource {
+    fun getMovieData(
+        query: String,
+        success: (NaverMovieResponse) -> Unit,
+        fail: (Throwable) -> Unit
+    )
+}
