+++ handnew04/model/src/main/java/com/example/model/di/LocalDataModule.kt
@@ -1,0 +1,9 @@
+package com.example.model.di
+
+import com.example.model.data.local.MovieLocalDataSource
+import com.example.model.data.local.MovieLocalDataSourceImpl
+import org.koin.dsl.module
+
+val localDataModule = module {
+    single<MovieLocalDataSource> { MovieLocalDataSourceImpl() }
+}
