+++ handnew04/model/src/main/java/com/example/model/di/NetworkManagerModule.kt
@@ -1,0 +1,8 @@
+package com.example.model.di
+
+import com.example.model.network.NetworkManager
+import org.koin.android.ext.koin.androidApplication
+import org.koin.dsl.module
+
+val networkManagerModule =
+    module { single { NetworkManager(androidApplication()) } }
