+++ handnew04/model/src/main/java/com/example/model/di/RepositoryModule.kt
@@ -1,0 +1,6 @@
+package com.example.model.di
+
+import com.example.model.data.MovieRepository
+import org.koin.dsl.module
+
+val movieRepositoryModule = module { single { MovieRepository(get(), get()) } }
