+++ handnew04/model/src/main/res/values/strings.xml
